BackEnd/API/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements-txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.1
    # via firecrawl-py
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   groq
    #   httpx
    #   sse-starlette
    #   starlette
    #   watchfiles
attrs==25.4.0
    # via aiohttp
beautifulsoup4==4.14.2
    # via gtr-task2
blockbuster==1.5.25
    # via langgraph-runtime-inmem
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   langgraph-cli
    #   rich-toolkit
    #   typer
    #   uvicorn
cloudpickle==3.1.1
    # via langgraph-api
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   uvicorn
cryptography==44.0.3
    # via langgraph-api
distro==1.9.0
    # via groq
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
fastapi==0.119.1
    # via gtr-task2
fastapi-cli==0.0.14
    # via fastapi
fastapi-cloud-cli==0.3.1
    # via fastapi-cli
firecrawl-py==4.5.0
    # via gtr-task2
forbiddenfruit==0.1.4 ; implementation_name == 'cpython'
    # via blockbuster
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
googleapis-common-protos==1.71.0
    # via opentelemetry-exporter-otlp-proto-http
greenlet==3.2.4 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'
    # via sqlalchemy
groq==0.33.0
    # via langchain-groq
grpcio==1.76.0
    # via
    #   grpcio-tools
    #   langgraph-api
grpcio-tools==1.76.0
    # via langgraph-api
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   firecrawl-py
    #   groq
    #   langgraph-api
    #   langgraph-sdk
    #   langsmith
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via opentelemetry-api
jinja2==3.1.6
    # via fastapi
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema-rs==0.29.1
    # via langgraph-api
langchain-core==1.0.0
    # via
    #   langchain-groq
    #   langgraph
    #   langgraph-api
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-groq==1.0.0
    # via gtr-task2
langgraph==1.0.1
    # via
    #   gtr-task2
    #   langgraph-api
    #   langgraph-runtime-inmem
langgraph-api==0.4.46
    # via langgraph-cli
langgraph-checkpoint==3.0.0
    # via
    #   langgraph
    #   langgraph-api
    #   langgraph-checkpoint-postgres
    #   langgraph-prebuilt
    #   langgraph-runtime-inmem
langgraph-checkpoint-postgres==3.0.0
    # via gtr-task2
langgraph-cli==0.4.4
    # via gtr-task2
langgraph-prebuilt==1.0.1
    # via langgraph
langgraph-runtime-inmem==0.14.1
    # via
    #   langgraph-api
    #   langgraph-cli
langgraph-sdk==0.2.9
    # via
    #   langgraph
    #   langgraph-api
    #   langgraph-cli
langsmith==0.4.37
    # via
    #   langchain-core
    #   langgraph-api
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
nest-asyncio==1.6.0
    # via firecrawl-py
opentelemetry-api==1.38.0
    # via
    #   langgraph-api
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp-proto-common==1.38.0
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-http==1.38.0
    # via langgraph-api
opentelemetry-proto==1.38.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.38.0
    # via
    #   langgraph-api
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.59b0
    # via opentelemetry-sdk
orjson==3.11.3
    # via
    #   langgraph-api
    #   langgraph-checkpoint-postgres
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.11.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.33.0
    # via
    #   googleapis-common-protos
    #   grpcio-tools
    #   langgraph-api
    #   opentelemetry-proto
psycopg==3.2.11
    # via langgraph-checkpoint-postgres
psycopg-pool==3.2.6
    # via langgraph-checkpoint-postgres
pycparser==2.23 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.12.3
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   firecrawl-py
    #   groq
    #   gtr-task2
    #   langchain-core
    #   langgraph
    #   langsmith
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.41.4
    # via pydantic
pydantic-settings==2.11.0
    # via gtr-task2
pygments==2.19.2
    # via rich
pyjwt==2.10.1
    # via langgraph-api
python-dotenv==1.1.1
    # via
    #   firecrawl-py
    #   langgraph-cli
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.3
    # via
    #   langchain-core
    #   uvicorn
requests==2.32.5
    # via
    #   firecrawl-py
    #   langsmith
    #   opentelemetry-exporter-otlp-proto-http
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rich==14.2.0
    # via
    #   gtr-task2
    #   rich-toolkit
    #   typer
rich-toolkit==0.15.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.7.1
    # via fastapi-cloud-cli
sentry-sdk==2.42.1
    # via fastapi-cloud-cli
setuptools==80.9.0
    # via grpcio-tools
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   groq
soupsieve==2.8
    # via beautifulsoup4
sqlalchemy==2.0.44
    # via sqlmodel
sqlmodel==0.0.27
    # via gtr-task2
sse-starlette==2.1.3
    # via
    #   langgraph-api
    #   langgraph-runtime-inmem
starlette==0.48.0
    # via
    #   fastapi
    #   langgraph-api
    #   langgraph-runtime-inmem
    #   sse-starlette
structlog==25.4.0
    # via
    #   langgraph-api
    #   langgraph-runtime-inmem
tenacity==9.1.2
    # via
    #   langchain-core
    #   langgraph-api
truststore==0.10.4
    # via langgraph-api
typer==0.20.0
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   groq
    #   grpcio
    #   langchain-core
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   psycopg
    #   psycopg-pool
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   starlette
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2 ; sys_platform == 'win32'
    # via psycopg
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
uvicorn==0.38.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
    #   langgraph-api
    #   sse-starlette
uvloop==0.22.1 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
    # via uvicorn
watchfiles==1.1.1
    # via
    #   langgraph-api
    #   uvicorn
websockets==15.0.1
    # via
    #   firecrawl-py
    #   uvicorn
xxhash==3.6.0
    # via langgraph
yarl==1.22.0
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
zstandard==0.25.0
    # via langsmith
